Unify charts behavior (partially)
This commit partially unifies charts behavior -- tree and circle
charts. There common behavior is now under control of the common code.
