+++ projects/GEDKeeper2/GKUI/Charts/CircleChart.cs
@@ -649,22 +649,6 @@ namespace GKUI.Charts
 
         #region Protected inherited methods
 
-        protected override bool IsInputKey(Keys keyData)
-        {
-            switch (keyData) {
-                case Keys.Add:
-                case Keys.Subtract:
-                case Keys.Left:
-                case Keys.Right:
-                case Keys.Up:
-                case Keys.Down:
-                case Keys.Back:
-                    return true;
-            }
-
-            return base.IsInputKey(keyData);
-        }
-
         protected override void OnPaint(PaintEventArgs e)
         {
             Render(e.Graphics, RenderTarget.rtScreen);
@@ -716,98 +700,27 @@ namespace GKUI.Charts
                     }
                     break;
                 }
-                case Keys.Left:
-                {
-                    HorizontalScroll.Value =
-                        Math.Max(HorizontalScroll.Value - HorizontalScroll.SmallChange, 0);
-                    PerformLayout();
-                    break;
-                }
-                case Keys.Right:
-                {
-                    HorizontalScroll.Value += HorizontalScroll.SmallChange;
-                    PerformLayout();
-                    break;
-                }
-                case Keys.Up:
-                {
-                    VerticalScroll.Value =
-                        Math.Max(VerticalScroll.Value - VerticalScroll.SmallChange, 0);
-                    PerformLayout();
-                    break;
-                }
-                case Keys.Down:
-                {
-                    VerticalScroll.Value += VerticalScroll.SmallChange;
-                    PerformLayout();
-                    break;
-                }
-                case Keys.PageUp:
-                {
-                    if (Keys.None == ModifierKeys) {
-                        VerticalScroll.Value =
-                            Math.Max(VerticalScroll.Value - VerticalScroll.LargeChange, 0);
-                    } else if (Keys.Shift == ModifierKeys) {
-                        HorizontalScroll.Value =
-                            Math.Max(HorizontalScroll.Value - HorizontalScroll.LargeChange, 0);
-                    }
-                    PerformLayout();
-                    break;
-                }
-                case Keys.PageDown:
-                {
-                    if (Keys.None == ModifierKeys) {
-                        VerticalScroll.Value += VerticalScroll.LargeChange;
-                    } else if (Keys.Shift == ModifierKeys) {
-                        HorizontalScroll.Value += HorizontalScroll.LargeChange;
-                    }
-                    PerformLayout();
-                    break;
-                }
-                case Keys.Home:
-                {
-                    if (Keys.None == ModifierKeys) {
-                        VerticalScroll.Value = 0;
-                    } else if (Keys.Shift == ModifierKeys) {
-                        HorizontalScroll.Value = 0;
-                    }
-                    PerformLayout();
-                    break;
-                }
-                case Keys.End:
-                {
-                    if (Keys.None == ModifierKeys) {
-                        VerticalScroll.Value = VerticalScroll.Maximum;
-                    } else if (Keys.Shift == ModifierKeys) {
-                        HorizontalScroll.Value = HorizontalScroll.Maximum;
-                    }
-                    PerformLayout();
-                    break;
-                }
-
-                case Keys.Back:
-                    NavPrev();
-                    return;
-
                 default:
+                {
                     base.OnKeyDown(e);
                     break;
+                }
             }
         }
 
         protected override void OnMouseDown(MouseEventArgs e)
         {
-            base.OnMouseDown(e);
             if ((MouseButtons.Left == e.Button) && (HorizontalScroll.Visible || VerticalScroll.Visible)) {
                 fMouseCaptured = 1;
                 fMouseCaptureX = e.X;
                 fMouseCaptureY = e.Y;
+            } else {
+                base.OnMouseDown(e);
             }
         }
 
         protected override void OnMouseUp(MouseEventArgs e)
         {
-            base.OnMouseUp(e);
             if (MouseButtons.Left == e.Button) {
                 if (2 > fMouseCaptured) {
                     CircleSegment selected = FindSegment(e.X, e.Y);
@@ -817,16 +730,13 @@ namespace GKUI.Charts
                 }
                 fMouseCaptured = 0;
                 Cursor = Cursors.Default;
-            } else if (MouseButtons.XButton1 == e.Button) {
-                NavPrev();
-            } else if (MouseButtons.XButton2 == e.Button) {
-                NavNext();
+            } else {
+                base.OnMouseUp(e);
             }
         }
 
         protected override void OnMouseMove(MouseEventArgs e)
         {
-            base.OnMouseMove(e);
             if (MouseButtons.Left == e.Button) {
                 if (0 == fMouseCaptured) {
                     CircleSegment selected = FindSegment(e.X, e.Y);
@@ -858,6 +768,8 @@ namespace GKUI.Charts
                     fMouseCaptureY = e.Y;
                     Cursor = Cursors.SizeAll;
                 }
+            } else {
+                base.OnMouseMove(e);
             }
         }
